feat: migrated modules to ml and adjusted unit tests and notebooks
